// src/api/back/models/GameServerDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Dota2Version,
    Dota2VersionFromJSON,
    Dota2VersionFromJSONTyped,
    Dota2VersionToJSON,
} from './';

/**
 * 
 * @export
 * @interface GameServerDto
 */
export interface GameServerDto {
    /**
     * 
     * @type {Dota2Version}
     * @memberof GameServerDto
     */
    version: Dota2Version;
    /**
     * 
     * @type {string}
     * @memberof GameServerDto
     */
    url: string;
}

export function GameServerDtoFromJSON(json: any): GameServerDto {
    return GameServerDtoFromJSONTyped(json, false);
}

export function GameServerDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameServerDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'version': Dota2VersionFromJSON(json['version']),
        'url': json['url'],
    };
}

export function GameServerDtoToJSON(value?: GameServerDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'version': Dota2VersionToJSON(value.version),
        'url': value.url,
    };
}