// src/api/back/models/GameSessionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    MatchInfoDto,
    MatchInfoDtoFromJSON,
    MatchInfoDtoFromJSONTyped,
    MatchInfoDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface GameSessionDto
 */
export interface GameSessionDto {
    /**
     * 
     * @type {string}
     * @memberof GameSessionDto
     */
    url: string;
    /**
     * 
     * @type {number}
     * @memberof GameSessionDto
     */
    matchId: number;
    /**
     * 
     * @type {MatchInfoDto}
     * @memberof GameSessionDto
     */
    info: MatchInfoDto;
}

export function GameSessionDtoFromJSON(json: any): GameSessionDto {
    return GameSessionDtoFromJSONTyped(json, false);
}

export function GameSessionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameSessionDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'url': json['url'],
        'matchId': json['matchId'],
        'info': MatchInfoDtoFromJSON(json['info']),
    };
}

export function GameSessionDtoToJSON(value?: GameSessionDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'url': value.url,
        'matchId': value.matchId,
        'info': MatchInfoDtoToJSON(value.info),
    };
}